+++ src/main/java/io/github/ImpactDevelopment/installer/Args.java
@@ -66,7 +66,7 @@ public class Args {
     @Parameter(names = {"--mc-dir", "--minecraft-dir", "--minecraft-directory", "--mc-path"}, description = "Path to the Minecraft directory")
     public String mcPath;
 
-    @Parameter(names = {"--optifine", "--of"}, description = "OptiFine, in the format like 1.12.2_HD_U_E2")
+    @Parameter(names = {"--optifine", "--of"}, description = "Path to an OptiFine installer jar")
     public String optifine;
 
     @Parameter(names = {"--no-ga", "--no-analytics", "--dnt", "--no-tracky"}, description = "Disable Google Analytics")
@@ -127,7 +127,7 @@ public class Args {
                 setImpactVersion(config, true, version);
                 try {
                     System.out.println(config.execute());
-                } catch (IOException e) {
+                } catch (Throwable e) {
                     throw new RuntimeException(e);
                 }
             }
@@ -144,7 +144,8 @@ public class Args {
             setImpactVersion(config, false, new ImpactVersionDisk(Paths.get(file)));
         }
         if (optifine != null) {
-            if (!config.setSettingValue(OptiFineSetting.INSTANCE, optifine)) {
+            config.setSettingValue(OptiFineToggleSetting.INSTANCE, true);
+            if (!config.setSettingValue(OptiFineFileSetting.INSTANCE, Paths.get(optifine))) {
                 throw new IllegalArgumentException(optifine + " is not found");
             }
         }