+++ src/main/java/io/github/ImpactDevelopment/installer/utils/OperatingSystem.java
@@ -22,6 +22,9 @@
 
 package io.github.ImpactDevelopment.installer.utils;
 
+import java.nio.file.Path;
+import java.nio.file.Paths;
+
 import static java.util.Locale.ROOT;
 
 /**
@@ -47,5 +50,17 @@ public enum OperatingSystem {
             return LINUX;
         }
         return UNKNOWN;
+    }
+
+    public static Path getDownloads() {
+        if (getOS() == OperatingSystem.LINUX) {
+            String xdg = System.getenv("XDG_DOWNLOAD_DIR");
+            if (!xdg.isEmpty()) return Paths.get(xdg);
+        }
+        return getHome().resolve("Downloads");
+    }
+
+    public static Path getHome() {
+        return Paths.get(System.getProperty("user.home"));
     }
 }